+++ lib/src/utils/utils.dart
@@ -1,0 +1,4 @@
+import 'package:azbox/src/widgets/language.dart';
+import 'package:flutter/material.dart';
+
+typedef ItemBuilder = Widget Function(Language language);