+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+* Added Dropdown Language Selector to select between the supported locales.
+* capitalize option when use translate()
+
 ## 1.0.1
 
 * supportedLocales is no longer needed since it is obtained from the project languages.